added router and page for product detail
