+++ client/src/pages/ProductDetail.js
@@ -1,0 +1,9 @@
+import React from "react";
+
+export function ProductDetail() {
+    return (
+    <div>
+        <p>Product Detail</p>
+    </div>
+    );
+}
